[브루니] 댓글 작성 / 삭제 구현

+++ src/main/java/kr/codesqaud/cafe/service/ArticleService.java
@@ -1,52 +1,60 @@
 package kr.codesqaud.cafe.service;
 
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import kr.codesqaud.cafe.controller.dto.ArticleParam;
+import kr.codesqaud.cafe.controller.dto.ArticleDetails;
+import kr.codesqaud.cafe.controller.dto.ArticleResponse;
+import kr.codesqaud.cafe.controller.dto.ArticleWithCommentCount;
+import kr.codesqaud.cafe.controller.dto.CommentResponse;
 import kr.codesqaud.cafe.controller.dto.req.ArticleEditRequest;
 import kr.codesqaud.cafe.controller.dto.req.PostingRequest;
 import kr.codesqaud.cafe.domain.article.Article;
+import kr.codesqaud.cafe.domain.comment.Comment;
+import kr.codesqaud.cafe.exception.InvalidOperationException;
 import kr.codesqaud.cafe.exception.NoAuthorizationException;
 import kr.codesqaud.cafe.exception.NotFoundException;
 import kr.codesqaud.cafe.repository.ArticleRepository;
+import kr.codesqaud.cafe.repository.CommentRepository;
 
 @Transactional(readOnly = true)
 @Service
 public class ArticleService {
 
 	private final ArticleRepository articleRepository;
+	private final CommentRepository commentRepository;
 
-	public ArticleService(ArticleRepository articleRepository) {
+	public ArticleService(ArticleRepository articleRepository, CommentRepository commentRepository) {
 		this.articleRepository = articleRepository;
+		this.commentRepository = commentRepository;
 	}
 
 	@Transactional
 	public void post(final PostingRequest request, final String userId) {
-		articleRepository.save(
-			new Article(null, userId, request.getTitle(), request.getContents(), LocalDateTime.now()));
+		articleRepository.save(Article.of(userId, request.getTitle(), request.getContents()));
 	}
 
-	public List<ArticleParam> getArticles() {
-		return articleRepository.findAll()
-			.stream()
-			.map(ArticleParam::from)
-			.collect(Collectors.toUnmodifiableList());
-	}
-
-	public ArticleParam findById(final Long id) {
-		return articleRepository.findById(id)
-			.map(ArticleParam::from)
+	public ArticleDetails getArticleDetails(final Long id) {
+		Article article = articleRepository.findById(id)
 			.orElseThrow(() -> new NotFoundException(String.format("%d번 게시글을 찾을 수 없습니다.", id)));
+		List<Comment> comments = commentRepository.findAllByArticleId(id);
+		return new ArticleDetails(ArticleResponse.from(article), comments.stream()
+			.map(CommentResponse::from)
+			.collect(Collectors.toUnmodifiableList()));
+	}
+
+	public List<ArticleWithCommentCount> getArticlesWithCommentCount() {
+		return articleRepository.findAllArticleWithCommentCount()
+			.stream()
+			.collect(Collectors.toUnmodifiableList());
 	}
 
 	public void validateHasAuthorization(final Long articleId, final String userId) {
 		articleRepository.findById(articleId)
-			.filter(article -> article.getWriter().equals(userId))
+			.filter(article -> article.isSameWriter(userId))
 			.orElseThrow(NoAuthorizationException::new);
 	}
 
@@ -63,6 +71,10 @@ public class ArticleService {
 	public void deleteArticle(final Long articleId) {
 		articleRepository.findById(articleId)
 			.orElseThrow(() -> new NotFoundException(String.format("%d번 게시글을 찾을 수 없습니다.", articleId)));
-		articleRepository.deleteById(articleId);
+		if (articleRepository.isPossibleDeleteById(articleId)) {
+			articleRepository.deleteById(articleId);
+			return;
+		}
+		throw new InvalidOperationException(articleId);
 	}
 }
